+++ packages/envd/internal/logs/exporter/mmds.go
@@ -41,7 +41,7 @@ func (opts *opts) addOptsToJSON(jsonLogs []byte) ([]byte, error) {
 	return data, err
 }
 
-func (w *HTTPLogsExporter) getMMDSToken(ctx context.Context) (string, error) {
+func (w *HTTPExporter) getMMDSToken(ctx context.Context) (string, error) {
 	request, err := http.NewRequestWithContext(ctx, http.MethodPut, "http://"+mmdsDefaultAddress+"/latest/api/token", new(bytes.Buffer))
 	if err != nil {
 		return "", err
@@ -69,7 +69,7 @@ func (w *HTTPLogsExporter) getMMDSToken(ctx context.Context) (string, error) {
 	return token, nil
 }
 
-func (w *HTTPLogsExporter) getMMDSOpts(ctx context.Context, token string) (*opts, error) {
+func (w *HTTPExporter) getMMDSOpts(ctx context.Context, token string) (*opts, error) {
 	request, err := http.NewRequestWithContext(ctx, http.MethodGet, "http://"+mmdsDefaultAddress, new(bytes.Buffer))
 	if err != nil {
 		return nil, err
